Alternative for long galleries: manifest file.
Instead of listing the gallery items in the frontmatter,
it can have a single key `manifest`, with its value a Zola-path pointing to a
data file. This file can be in any supported textual format.

Zola-paths are resolved from content directory. So if a gallery manifest is in
the `w/` folder next to a wrestler's article, the path is `"@/w/some-file.toml"`.

As a testbed, converted Justin Joy's gallery.