Add a small animation to indicate that search index is still loading.
To test: set network throttling in devtools to e.g. 3G.
Notice that search is not functional after focusing and typing, and that two
requests were fired to fetch the minisearch code and the site index (this part
is not new). While the two are pending, the magnifying glass icon in the search
field is replaced with a loader, which slowly spins.